requirements.txt
Django==4.2.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
Pillow==11.2.1
